Separate CR and MTR debug pages
